fund types fuzzing
