+++ user-fund-storage/fuzz/.gitignore
@@ -1,0 +1,4 @@
+
+target
+corpus
+artifacts
