+++ user-fund-storage/src/types/mod.rs
@@ -1,5 +1,7 @@
 pub mod fund_item;
+pub mod fund_list_info;
 pub mod fund_type;
 
 pub use fund_item::*;
+pub use fund_list_info::*;
 pub use fund_type::*;
